Move the HA basic HTTP API to the api namespace (to match the WS API)
